Punted private functions into individual static classes
And grouped associated classes with it
