--- api/ExpressedRealms.Expressions.API/ExpressionEndpoints/DTOs/PublishTypeDto.cs
@@ -1,3 +1,0 @@
-namespace ExpressedRealms.Expressions.API.ExpressionEndpoints.DTOs;
-
-public record PublishTypeDto(int Id, string Name);
