+++ api/ExpressedRealms.Expressions.API/ExpressionEndpoints/GetEditExpression/PublishTypeDto.cs
@@ -1,0 +1,3 @@
+namespace ExpressedRealms.Expressions.API.ExpressionEndpoints.GetEditExpression;
+
+public record PublishTypeDto(int Id, string Name);
