+++ api/ExpressedRealms.Expressions.API/ExpressionEndpoints/UpdateHierarchy/EditExpressionHierarchyRequest.cs
@@ -1,4 +1,4 @@
-namespace ExpressedRealms.Expressions.API.ExpressionEndpoints.Requests;
+namespace ExpressedRealms.Expressions.API.ExpressionEndpoints.UpdateHierarchy;
 
 public class EditExpressionHierarchyItemRequest
 {
